Extract waitForElementAnimationToFinish to new Conditions class
Starts extraction of Conditions into a separate class to enable re-use.
